# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "api/send-email": {
            "post": {
                "tags": [
                    "Add Email"
                ],
                "summary": "User Send Email for add",
                "description": "User Send Email for add",
                "operationId": "User Send Email for add",
                "requestBody": {
                    "description": "User Send Email for add",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "required"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "test@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/add-email": {
            "post": {
                "tags": [
                    "Add Email"
                ],
                "summary": "User Send code for add",
                "description": "User Send code for add",
                "operationId": "User Send code for add",
                "requestBody": {
                    "description": "User Send code for add",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "required"
                                ],
                                "properties": {
                                    "random_int": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "123456"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "8"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/send-number": {
            "post": {
                "tags": [
                    "Add Number"
                ],
                "summary": "User Send Number for add",
                "description": "User Send Number for add",
                "operationId": "User Send Number for add",
                "requestBody": {
                    "description": "User Send Number for add",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "required"
                                ],
                                "properties": {
                                    "number": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "1234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/add-number": {
            "post": {
                "tags": [
                    "Add Number"
                ],
                "summary": "User Send code for add",
                "description": "User Send code for add",
                "operationId": "User Send Code for add",
                "requestBody": {
                    "description": "User Send code for add",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "required"
                                ],
                                "properties": {
                                    "random_int": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "1234567"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "8"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "api/code-sending": {
            "post": {
                "tags": [
                    "Forgot-password"
                ],
                "summary": "User Send Email",
                "description": "User Send Email",
                "operationId": "User Send Email",
                "requestBody": {
                    "description": "User Send Email",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "required"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "test@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "api/update-number": {
            "post": {
                "tags": [
                    "Change Number"
                ],
                "summary": "User Send code for change",
                "description": "User Send code for change",
                "operationId": "User Send code ",
                "requestBody": {
                    "description": "User Send code for change",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "required"
                                ],
                                "properties": {
                                    "random_int": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "123456"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "8"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}